fix: remove realtime mixin
