+++ .eslintignore
@@ -1,3 +1,4 @@
 build/*
 __mocks__/*
 node_modules/*
+src/types/*.d.ts
