+++ package.json
@@ -12,9 +12,9 @@
     "./types": "./types/index.d.ts"
   },
   "dependencies": {
-    "@geoblocks/ol-maplibre-layer": "^1.0.0",
-    "@turf/helpers": "7.0.0",
-    "@turf/transform-rotate": "7.0.0",
+    "@geoblocks/ol-maplibre-layer": "^1.0.1",
+    "@turf/helpers": "7.1.0",
+    "@turf/transform-rotate": "7.1.0",
     "lodash.debounce": "4.0.8",
     "lodash.throttle": "4.1.1",
     "uuid": "10.0.0"
@@ -24,22 +24,23 @@
     "ol": ">=7"
   },
   "devDependencies": {
-    "@babel/preset-env": "^7.24.7",
+    "@babel/preset-env": "^7.25.4",
     "@babel/preset-typescript": "^7.24.7",
-    "@commitlint/cli": "19.3.0",
-    "@commitlint/config-conventional": "19.2.2",
-    "@geops/eslint-config-react": "^1.0.2",
+    "@commitlint/cli": "19.5.0",
+    "@commitlint/config-conventional": "19.5.0",
+    "@geops/eslint-config-react": "^1.1.1",
     "@types/geojson": "7946.0.14",
+    "@types/lodash": "^4.17.7",
     "@types/lodash.debounce": "4.0.9",
     "@types/lodash.throttle": "4.1.9",
-    "@types/mapbox-gl": "3.1.0",
+    "@types/mapbox-gl": "3.4.0",
     "@types/offscreencanvas": "2019.7.3",
     "@types/topojson": "3.2.6",
     "@types/uuid": "10.0.0",
-    "@typescript-eslint/eslint-plugin": "7.14.1",
-    "@typescript-eslint/parser": "7.14.1",
-    "cypress": "13.12.0",
-    "esbuild": "0.21.5",
+    "@typescript-eslint/eslint-plugin": "8.5.0",
+    "@typescript-eslint/parser": "8.5.0",
+    "cypress": "13.14.2",
+    "esbuild": "0.23.1",
     "esdoc": "1.1.0",
     "esdoc-ecmascript-proposal-plugin": "1.0.0",
     "esdoc-publish-html-plugin": "1.1.2",
@@ -49,13 +50,13 @@
     "eslint-config-airbnb": "19.0.4",
     "eslint-config-airbnb-typescript": "18.0.0",
     "eslint-config-prettier": "9.1.0",
-    "eslint-plugin-cypress": "3.3.0",
-    "eslint-plugin-import": "2.29.1",
-    "eslint-plugin-jsx-a11y": "6.9.0",
-    "eslint-plugin-prettier": "5.1.3",
-    "eslint-plugin-react": "7.34.3",
+    "eslint-plugin-cypress": "3.5.0",
+    "eslint-plugin-import": "2.30.0",
+    "eslint-plugin-jsx-a11y": "6.10.0",
+    "eslint-plugin-prettier": "5.2.1",
+    "eslint-plugin-react": "7.36.1",
     "fixpack": "4.0.0",
-    "husky": "9.0.11",
+    "husky": "9.1.6",
     "is-ci": "3.0.1",
     "jest": "29.7.0",
     "jest-canvas-mock": "2.5.2",
@@ -64,23 +65,23 @@
     "jest-serializer-html": "7.1.0",
     "jest-transformer-svg": "2.0.2",
     "jest-websocket-mock": "2.5.0",
-    "lint-staged": "15.2.7",
-    "maplibre-gl": "4.5.0",
+    "lint-staged": "15.2.10",
+    "maplibre-gl": "4.7.0",
     "mock-socket": "9.3.1",
-    "next": "14.2.4",
+    "next": "14.2.11",
     "next-transpile-modules": "10.0.1",
-    "ol": "9.2.4",
+    "ol": "10.1.0",
     "openapi-typescript": "6.7.5",
-    "prettier": "3.3.2",
+    "prettier": "3.3.3",
     "raw-loader": "4.0.2",
     "sort-json": "2.0.1",
     "standard-version": "9.5.0",
-    "start-server-and-test": "2.0.4",
-    "stylelint": "16.6.1",
-    "stylelint-config-recommended-scss": "14.0.0",
+    "start-server-and-test": "2.0.7",
+    "stylelint": "16.9.0",
+    "stylelint-config-recommended-scss": "14.1.0",
     "stylelint-config-standard": "36.0.1",
-    "stylelint-scss": "6.3.2",
-    "typescript": "5.5.2"
+    "stylelint-scss": "6.6.0",
+    "typescript": "5.6.2"
   },
   "scripts": {
     "apidoc": "esdoc && cp apidoc/index.json doc/src/components/Esdoc",
